Rename the app to BC Parks Foundation
