+++ client/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.bpfall2019",
+    package = "com.bcparksfoundation",
 )
 
 android_resource(
     name = "res",
-    package = "com.bpfall2019",
+    package = "com.bcparksfoundation",
     res = "src/main/res",
 )
