+++ client/android/app/src/main/java/com/bcparksfoundation/MainActivity.java
@@ -1,4 +1,4 @@
-package com.bpfall2019;
+package com.bcparksfoundation;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "BPFall2019";
+    return "BCParksFoundation";
   }
 }
