+++ client/android/app/src/main/java/com/bcparksfoundation/MainApplication.java
@@ -1,4 +1,4 @@
-package com.bpfall2019;
+package com.bcparksfoundation;
 
 import android.app.Application;
 import android.content.Context;
