+++ client/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BPFall2019</string>
+    <string name="app_name">BC Parks Foundation</string>
 </resources>
