+++ client/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "BPFall2019",
-  "displayName": "BPFall2019"
+  "name": "BCParksFoundation",
+  "displayName": "BC Parks Foundation"
 }
