+++ client/ios/BCParksFoundation/Images.xcassets/AppIcon.appiconset/Contents.json
