+++ client/ios/BCParksFoundationTests/BCParksFoundationTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface BPFall2019Tests : XCTestCase
+@interface BCParksFoundationTests : XCTestCase
 
 @end
 
-@implementation BPFall2019Tests
+@implementation BCParksFoundationTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
