+++ client/ios/BCParksFoundationTests/Info.plist
