+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "BPFall2019",
+  "name": "BCParksFoundation",
   "version": "0.0.1",
   "private": true,
   "scripts": {
